Applying Creational Design Pattern - Builder Pattern - Part 2 (Updates)
feat: added new class Director to execute the building steps (booking of an appointment) in a particular sequence. Initialized it's constructor as well. Also included 4 methods for the building sequence, one for booking a basic appointment which requires just the patient, country and city and the others for booking appointments with specialised arguments. (2) Implemented the function setPatient in class AppointmentBuilder1 to set the patient in which the appointment will be booked for.